Changes to oneRingAlgoTest (actually 2 ring now)

+++ src/main/java/cdio/drone/interfaces/IDroneCommander.java
@@ -28,7 +28,8 @@ public interface IDroneCommander {
     void flyLeft(int distanceMilli) throws DroneCommanderException;
     void flyRight(int distanceMilli) throws DroneCommanderException;
     void flyUpToAltitude(int altitude);
-    void flyDownToAltitude(int altitude);
+
+    void flyDownToAltitude(float altitude);
 
     void circleAroundObjectV2();
